% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hit_factors.R
\name{plot_hit_factors}
\alias{plot_hit_factors}
\title{plot_hit_factors}
\usage{
plot_hit_factors(data)
}
\arguments{
\item{data}{A dataframe containing song audio features.}
}
\value{
A ggplot bar chart.
}
\description{
Creates a bar plot showing feature contributions to the hit score.
}
\examples{
test_data <- data.frame(
    energy = c(0.8, 0.6),
  danceability = c(0.7, 0.5),
  valence = c(0.6, 0.4),
  popularity = c(80, 50),
  loudness = c(-5, -10),
  tempo = c(120, 100),
  speechiness = c(0.05, 0.1),
  instrumentalness = c(0.0, 0.2),
  explicit = c(TRUE, FALSE)
)
plot_hit_factors(test_data)
}
